pyproject.toml
[tool.poetry]
name = "ekklesia-voting"
version = "23.06.0"
description = "Pseudonymous voting for the Ekklesia e-democracy platform"
keywords = ["ekklesia", "edemocracy", "evoting", "voting", "participation"]
readme = "README.md"
license = "AGPL-3.0-or-later"
authors = ["Tobias dpausp <dpausp@posteo.de>"]
classifiers = [
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11"
]


[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.4.2"
dataclasses-json = "^0.4.3"
psycopg2 = "^2.8.5"
sqlalchemy = "^1.4.47"
ekklesia-common = { git = "https://github.com/edemocracy/ekklesia-common", rev = "master" }
base32-crockford = "^0.3.0"
babel = "^2.9.0"
deform = "^2.0.15"
eliot-tree = "^19.0.1"
pdbpp = "^0.10.2"
Jinja2 = "^3.1"
typer = "^0.7.0"
gunicorn = "^20.1.0"
ipython = "^8.11.0"


[tool.poetry.dev-dependencies]
colorama = "^0.4.3"
mimesis-factory = "^1.1.0"
pytest = "^6.0.1"
pytest-cov = "^2.8.1"
pytest-factoryboy = "^2.0.3"
factory_boy = { git = "https://github.com/FactoryBoy/factory_boy", rev = "master" }
pytest-instafail = "^0.4.1"
pytest-localserver = "^0.5.0"
pytest-mock = "^3.1.0"
pytest-pspec = "^0.0.4"
responses = "^0.10.14"
typing-extensions = "^3.7.4"
webtest = "^3.0.0"
doit = "^0.36.0"
multiprocess = "^0.70.12"


[tool.coverage.report]
skip_covered = true
# Regexes for lines to exclude from consideration
exclude_lines = [
    "pragma: no cover",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:"
]


[tool.coverage.run]
source = [
    "ekklesia_voting"
]
omit = [
    "src/ekklesia_voting/runserver.py",
    "src/ekklesia_voting/concepts/*/__init__.py",
]

[tool.pylint.basic]
good-names = "q, s, f, wf, m, k, v, id"

[tool.pylint.imports] # Taken care of by isort, mostly included for error messages
known-third-party = "ekklesia_common"


[tool.pylint.master]
disable = "C0114, C0115, C0116, R0903"


[tool.pylint.format]
max-line-length = "88"
ignore-long-lines = "^\\s*(# )?<?https?://\\S+>?$"


[tool.pytest.ini_options]
addopts = """\
  --show-capture=no \
  -ra \
  --tb=short \
  -k "not integration" \
  --strict-markers \
"""
filterwarnings = [
    "ignore::DeprecationWarning"
]
norecursedirs = "tests/helpers"
testpaths = "tests"
markers = [
    "integration: An Integration test which is allowed to modify the test database."
]


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
